+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log (@egomobile/http-server)
 
+## 0.50.0
+
+- **BREAKING CHANGE**: [@Describe() decorator][https://egomobile.github.io/node-http-server/functions/describe.html] will throw an exception, if it is tried to be used more than once in a controller class
+- improve constructor arguments of [@Describe() decorator][https://egomobile.github.io/node-http-server/functions/describe.html]
+
 ## 0.49.1
 
 - **BREAKING CHANGE**: tests are sorted now the following way: